feat(FOROME-791): refactoring

+++ src/pages/filter/ui/filter-control/filter-control-refiner/filter-control-refiner.store.tsx
@@ -40,6 +40,12 @@ class FilterControlRefinerStore {
     const { conditions } = filterStore
 
     if (filterStore.actionName === ActionFilterEnum.Delete) {
+      if (!datasetStore.activePreset) {
+        showToast(t('error.choosePresetFirst'), 'error')
+
+        return
+      }
+
       presetStore.deletePreset()
       filterStore.resetSelectedFilters()
       datasetStore.fetchDsStatAsync()
